backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python
pillow
starlette
